# apps/docs/src/openapi/jurnal-zitn-api.json
{
...
  "info": {
    "description": "HTTP API for Jurnal ZITN — the open-source trading journal.\n\nThis reference is generated from the `@swagger` JSDoc blocks that live next to each route in the API source, so it always matches the endpoints the app actually serves. All routes are mounted under `/api`. The same API backs both the hosted app (`api.tradr.cloud`) and a self-hosted instance (reached through the web container at `/api`). Most endpoints require an authenticated session cookie.",
    "title": "Jurnal ZITN API",
    "version": "0.15.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/admin/feedback": {
      "get": {
        "description": "Admin-gated (403 `ADMIN_REQUIRED`) and per-user rate limited. Filters are optional and combine with AND: `type` (bug/feature/general/question), `status` (baru/ditinjau/direncanakan/selesai/ditolak) and `source` (jurnal/lembar). Cursor-paginated over `(created_at, id)` descending; `cursor` is the opaque base64 cursor from a prior page, absent/invalid ⇒ first page. `limit` defaults to 25, clamped to [1, 100]. Response `{ items, nextCursor }`; each item carries `userEmail` (PII) because this is the triage inbox.\n",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "schema": {
              "enum": [
                "bug",
                "feature",
                "general",
                "question"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "enum": [
                "baru",
                "ditinjau",
                "direncanakan",
                "selesai",
                "ditolak"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "schema": {
              "enum": [
                "jurnal",
                "lembar"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items: Feedback[], nextCursor: string | null }."
          },
          "400": {
            "description": "VALIDATION_ERROR — malformed filter or limit."
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "ADMIN_REQUIRED — authenticated but not an admin."
          },
          "429": {
            "description": "Admin rate limit reached (60 / 60 s per user)."
          }
        },
        "summary": "List in-app feedback, newest-first (admin only).",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/feedback/{id}": {
      "patch": {
        "description": "Admin-gated (403 `ADMIN_REQUIRED`). Body `{ status }`; PATCH (never GET) per the SameSite=Lax CSRF posture. Returns the updated row. A non-UUID id is a 400 `VALIDATION_ERROR`; an unknown id is a 404 `NOT_FOUND`.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "status": {
                    "enum": [
                      "baru",
                      "ditinjau",
                      "direncanakan",
                      "selesai",
                      "ditolak"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated feedback row."
          },
          "400": {
            "description": "VALIDATION_ERROR — id is not a UUID or the status is invalid."
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "ADMIN_REQUIRED — authenticated but not an admin."
          },
          "404": {
            "description": "NOT_FOUND — no such feedback row."
          },
          "429": {
            "description": "Admin rate limit reached (60 / 60 s per user)."
          }
        },
        "summary": "Move one feedback row to another triage status (admin only).",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/stats": {
      "get": {
        "description": "Admin-gated (403 `ADMIN_REQUIRED` for non-admins) and per-user rate limited (60 / 60 s). Returns total users, \"Active now\" users (still-valid sessions in the last 30 min — `activeUsersWindowMinutes` is serialized for honest labeling), positions by status, and revenue as micro-USD integer strings: `allTime` and the reversal-attributed UTC `currentMonth`, with the pinned `basis: 'purchased-credit-volume'`. An empty instance returns well-formed zeros, never errors.\n",
        "responses": {
          "200": {
            "description": "AdminStats — totals, activeUsers, positions, revenue."
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "ADMIN_REQUIRED — authenticated but not an admin."
          },
          "429": {
            "description": "Admin rate limit reached (60 / 60 s per user)."
          }
        },
        "summary": "Platform statistics (admin only).",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/usage": {
      "get": {
        "description": "Admin-gated (403 `ADMIN_REQUIRED`). `from`/`to` are optional ISO 8601 datetimes; the default window is the trailing 30 days ending now. `from > to` and ranges over 366 days are a 400 `VALIDATION_ERROR`. Returns `period`, `totals` (token/credit integer strings; `providerCost` is null when zero covered rows — pre-0013 rows have no recorded raw cost, see `providerCostCoverage`), the UTC-day `series`, `topUsers` (max 50 by billed credits), and the reversal-attributed `revenue { credited, reversed, net }`. Empty/future ranges return well-formed zeros.\n",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "AdminUsage — period, totals, series, topUsers, revenue."
          },
          "400": {
            "description": "VALIDATION_ERROR — malformed datetimes, from > to, or range > 366 days."
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "ADMIN_REQUIRED — authenticated but not an admin."
          },
          "429": {
            "description": "Admin rate limit reached (60 / 60 s per user)."
          }
        },
        "summary": "Platform usage and revenue over a period (admin only).",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "description": "Admin-gated (403 `ADMIN_REQUIRED`). Cursor-paginated over `(created_at, id)` descending — new signups on page one. `cursor` is the opaque base64 cursor from a prior page; absent/invalid ⇒ first page. `limit` defaults to 25, clamped to [1, 100]; a non-numeric limit is a 400 `VALIDATION_ERROR`. Response `{ items, nextCursor }`; `lastActiveAt` is the last recorded session activity (nullable, may be arbitrarily old); `emailVerified` is the stored verified flag — a read-only signal, nothing is gated on it in v1. No secret fields are returned.\n",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items: AdminUserListItem[], nextCursor: string | null }."
          },
          "400": {
            "description": "VALIDATION_ERROR — malformed limit."
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "ADMIN_REQUIRED — authenticated but not an admin."
          },
          "429": {
            "description": "Admin rate limit reached (60 / 60 s per user)."
          }
        },
        "summary": "List users, cursor-paginated newest-first (admin only).",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "description": "Admin-gated (403 `ADMIN_REQUIRED`). Returns the list fields plus `positionCount`, `advisorTurns` (platform-key advisor turns, current UTC month), all-time `usage` token/credit sums, and `walletBalance` ('0' when no wallet row) — all bigint sums as integer strings. A non-UUID id is a 400 `VALIDATION_ERROR`; an unknown id is a 404 `NOT_FOUND`.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "AdminUserDetail."
          },
          "400": {
            "description": "VALIDATION_ERROR — id is not a UUID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "ADMIN_REQUIRED — authenticated but not an admin."
          },
          "404": {
            "description": "NOT_FOUND — no such user."
          },
          "429": {
            "description": "Admin rate limit reached (60 / 60 s per user)."
          }
        },
        "summary": "Per-user admin detail (admin only).",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/admin/users/{id}/reset-preview": {
      "get": {
        "description": "Admin-gated (403 `ADMIN_REQUIRED`). Read-only. Returns the row counts a factory reset would remove, split into `tradingData` (always removed) and `settings` (removed only when the reset is requested with `removeSettings: true`). Both halves are always counted, so a client can show what including settings would add without a second request. Counts are a snapshot and are not a promise about a later reset — `POST .../reset` re-counts from its own deletes.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "AdminResetPreview — { userId, email, tradingData, settings }."
          },
          "400": {
            "description": "VALIDATION_ERROR — id is not a UUID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "ADMIN_REQUIRED — authenticated but not an admin."
          },
          "404": {
            "description": "NOT_FOUND — no such user."
          },
          "429": {
            "description": "Admin rate limit reached (60 / 60 s per user)."
          }
        },
        "summary": "What a factory reset would delete for this user (admin only).",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Hosted (tradr.cloud)",
      "url": "https://api.tradr.cloud"
    },
    {
      "description": "Local development (API PORT default)",
      "url": "http://localhost:3100"
    }
  ],
...
}